GH-723 code cleanup
