+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/Send/SendInputViewController.swift
@@ -21,11 +21,6 @@ public class SendInputViewController: UIViewController {
     @IBOutlet weak var accountBalanceHeaderView: AccountBalanceView!
     @IBOutlet weak var feeCalculationView: FeeCalculationView!
 
-    // Either transactionFeeView or tokenBalanceView and feeBalanceView are visible at the same time
-    @IBOutlet weak var transactionFeeView: TransactionFeeView!
-    @IBOutlet weak var tokenBalanceView: TransactionFeeView!
-    @IBOutlet weak var feeBalanceView: TransactionFeeView!
-
     weak var delegate: SendInputViewControllerDelegate?
 
     private var keyboardBehavior: KeyboardAvoidingBehavior!
@@ -75,15 +70,10 @@ public class SendInputViewController: UIViewController {
         tokenInput.tokenCode = model.tokenData.code
         tokenInput.delegate = self
         tokenInput.textInput.accessibilityIdentifier = "transaction.amount"
-        if tokenID == feeTokenID {
-            feeCalculationView.calculation = SendEthFeeCalculation()
-        } else {
-            feeCalculationView.calculation = SendERC20FeeCalculation()
-        }
+        feeCalculationView.calculation = tokenID == feeTokenID ? SendEthFeeCalculation() : SendERC20FeeCalculation()
         model.start()
-
         DispatchQueue.main.async {
-            // For unknown reasons, the blockies (identicon) does not show up if updated in the viewDidLoad
+            // For unknown reasons, the identicon does not show up if updated in the viewDidLoad
             self.accountBalanceHeaderView.address = ApplicationServiceRegistry.walletService.selectedWalletAddress
         }
     }
@@ -105,69 +95,32 @@ public class SendInputViewController: UIViewController {
 
     func updateFromViewModel() {
         accountBalanceHeaderView.amount = model.tokenData
-
-        let formatter = TokenNumberFormatter.ERC20Token(code: model.feeAmountTokenData.code,
-                                                        decimals: model.feeAmountTokenData.decimals,
-                                                        displayedDecimals: 5)
-
         if tokenID == feeTokenID {
             let calculation = feeCalculationView.calculation as! SendEthFeeCalculation
-
-
-            let fee = formatter.string(from: model.feeAmountTokenData.balance ?? 0)
-            calculation.networkFeeLine.set(value: fee)
-            let resultingBalance = formatter.string(from: model.feeResultingBalanceTokenData.balance ?? 0)
-            calculation.resultingBalanceLine.set(value: resultingBalance)
-
-            calculation.errorLine.text = ""
-            calculation.resultingBalanceLine.set(error: nil)
-            if model.hasEnoughFunds() == false {
-                let error = FeeCalculationError.insufficientBalance
-                calculation.errorLine.text = error.localizedDescription
-                calculation.resultingBalanceLine.set(error: error)
-            }
-            calculation.update()
-            feeCalculationView.update()
+            calculation.networkFeeLine.set(value: model.feeAmountTokenData)
+            calculation.resultingBalanceLine.set(value: model.feeResultingBalanceTokenData)
+            calculation.setBalanceError(feeBalanceError())
         } else {
             let calculation = feeCalculationView.calculation as! SendERC20FeeCalculation
-            formatter.tokenCode = model.resultingTokenData.code
-            formatter.decimals = model.resultingTokenData.decimals
-            let resultingBalance = formatter.string(from: model.resultingTokenData.balance ?? 0)
-            calculation.resultingBalanceLine.set(value: resultingBalance)
-
-            formatter.tokenCode = model.feeAmountTokenData.code
-            formatter.decimals = model.feeAmountTokenData.decimals
-
-            let feeBalance = formatter.string(from: model.feeBalanceTokenData.balance ?? 0)
-            calculation.networkFeeBalance.set(value: feeBalance)
-
-            let fee = formatter.string(from: model.feeAmountTokenData.balance ?? 0)
-            calculation.networkFeeLine.set(value: fee)
-
-            let feeResultingBalance = formatter.string(from: model.feeResultingBalanceTokenData.balance ?? 0)
-            calculation.networkFeeResultingBalanceLine.set(value: feeResultingBalance)
-
-            calculation.errorLine.text = ""
-            calculation.resultingBalanceLine.set(error: nil)
-            calculation.networkFeeBalanceErrorLine.text = ""
-            calculation.networkFeeResultingBalanceLine.set(error: nil)
-            if model.hasEnoughFunds() == false {
-                let error = FeeCalculationError.insufficientBalance
-
-                if (model.resultingTokenData.balance ?? 0) < 0 {
-                    calculation.errorLine.text = error.localizedDescription
-                    calculation.resultingBalanceLine.set(error: error)
-                }
-
-                if (model.feeResultingBalanceTokenData.balance ?? 0) < 0 {
-                    calculation.networkFeeBalanceErrorLine.text = error.localizedDescription
-                    calculation.networkFeeResultingBalanceLine.set(error: error)
-                }
-            }
-            calculation.update()
-            feeCalculationView.update()
+            calculation.resultingBalanceLine.set(value: model.resultingTokenData)
+            calculation.setBalanceError(tokenBalanceError())
+            calculation.networkFeeBalance.set(value: model.feeBalanceTokenData)
+            calculation.networkFeeLine.set(value: model.feeAmountTokenData)
+            calculation.networkFeeResultingBalanceLine.set(value: model.feeResultingBalanceTokenData)
+            calculation.setFeeBalanceError(feeBalanceError())
         }
+        feeCalculationView.update()
         nextBarButton.isEnabled = model.canProceedToSigning
+    }
+
+    private func tokenBalanceError() -> Error? {
+        let isNegativeBalance = (model.resultingTokenData.balance ?? 0) < 0
+        return  model.hasEnoughFunds() == false && isNegativeBalance ? FeeCalculationError.insufficientBalance : nil
+    }
+
+    private func feeBalanceError() -> Error? {
+        let isNegativeFeeBalance = (model.feeResultingBalanceTokenData.balance ?? 0) < 0
+        return model.hasEnoughFunds() == false && isNegativeFeeBalance ? FeeCalculationError.insufficientBalance : nil
     }
 
     @IBAction func proceedToSigning(_ sender: Any) {
@@ -222,75 +175,4 @@ extension SendInputViewController: VerifiableInputDelegate {
         keyboardBehavior.activeTextField = verifiableInput.textInput
     }
 
-}
-
-class SendEthFeeCalculation: FeeCalculation {
-
-    enum Strings {
-
-        static let currentBalance = LocalizedString("safe_balance", comment: "Current balance")
-        static let networkFee = LocalizedString("transaction_fee", comment: "Network fee")
-        static let resultingBalance = LocalizedString("balance_after_transfer", comment: "Balance after transfer")
-        static let loading = "-"
-        static let feeInfo = "[?]"
-
-    }
-
-    var networkFeeLine: FeeCalculationAssetLine
-    var resultingBalanceLine: FeeCalculationAssetLine
-    var errorLine: FeeCalculationErrorLine
-
-    required init() {
-        networkFeeLine = FeeCalculationAssetLine()
-            .set(name: Strings.networkFee)
-            .set(value: Strings.loading)
-            .set(button: Strings.feeInfo, target: nil, action: Selector(("showTransactionFeeInfo")))
-        resultingBalanceLine = FeeCalculationAssetLine()
-            .set(style: .balance)
-            .set(name: Strings.resultingBalance)
-            .set(value: Strings.loading)
-        errorLine = FeeCalculationErrorLine(text: Strings.loading)
-        super.init()
-        update()
-    }
-
-    func update() {
-        let section = FeeCalculationSection([networkFeeLine, resultingBalanceLine, errorLine])
-        section.border = nil
-        section.insets = .zero
-        elements = [section]
-    }
-
-}
-
-class SendERC20FeeCalculation: SendEthFeeCalculation {
-
-    var networkFeeBalance: FeeCalculationAssetLine
-    var networkFeeResultingBalanceLine: FeeCalculationAssetLine
-    var networkFeeBalanceErrorLine: FeeCalculationErrorLine
-
-    required init() {
-        networkFeeBalance = FeeCalculationAssetLine()
-            .set(name: Strings.currentBalance)
-            .set(value: Strings.loading)
-        networkFeeResultingBalanceLine = FeeCalculationAssetLine()
-            .set(style: .balance)
-            .set(name: Strings.resultingBalance)
-            .set(value: Strings.loading)
-        networkFeeBalanceErrorLine = FeeCalculationErrorLine(text: Strings.loading)
-        super.init()
-    }
-
-    override func update() {
-        let section = FeeCalculationSection([resultingBalanceLine,
-                                             errorLine,
-                                             FeeCalculationSpacingLine(spacing: 20),
-                                             networkFeeBalance,
-                                             networkFeeLine,
-                                             networkFeeResultingBalanceLine,
-                                             networkFeeBalanceErrorLine])
-        section.border = nil
-        section.insets = .zero
-        elements = [section]
-    }
 }
